class-33/demo/blogapi-permissions/pyproject.toml
[tool.poetry]
name = "blogapi"
version = "0.1.0"
description = ""
authors = ["Ahmad Alawad <ahmad.alawad.sf@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.2.5"
djangorestframework = "^3.12.4"
djangorestframework_simplejwt = "^4.7.2"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
black = "^21.7b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
